add lazy operator for ranges, * and + (wip)
Now, we store a map of Id => hashset![index of lazy parents], which serves no
purposes, except see if there are multiple lazy operators for the same leaf.
It's only a theoretical possibility, and I'm not sure it has any practical
use cases; perhaps it can simply be forbidden.

The other problem is if a lazy operator is at the top, and another greedy
*, +, or range is below. The children of the greedy node will be tagged as
non-greedy, which must be avoided, but how?

Ex: ([a-c]+|[b-d]+)+?a on "abcda"