remove Boxes from LexGen and Lexer
I'll have to investigate how to avoid the transfer and duplication of those tables
in tests and in the generated source code.

If it's only generated code, the tables can be changed to array references, and the
maps can be initialized from arrays of tuples (there's no avoiding redundancy here).

For the tests and other dynamically-generated lexers, we'd need to store those
arrays somewhere where it lives longer than the lexer. It would be fine for tests,
but could otherwise be less than ideal. Perhaps an optional series of Vec stored
in the lexer and referenced?

It would be easier to create a Box or a Vec from either dynamic or static data,
somehow (and what about Drop?).