add lexer for regular expression parser
As a first step, we
- manually create the regular expression for the rlexer's lexicon
- add a simple custom top-down parser
- add the logic to convert AST to regular expression
- make a first version of the rlexer binary

Later, we can extend the lexicon from a proper file if necessary.
We can also use it to create the lexer for the rparser, and modify
the top-down parser of rlexer as a starting point for the first version
of the rparser's parser. After that, we can create a proper grammar
file for the rlexer and rparser tools.