simplify opcode logic
The part that swaps the exit when there's left recursion is commented out because it
generates all sorts of problems. A possibly easier solution is to handle the
stack_t items and to call the user's exit_NT(...) from the internal enter_NT(...)
methods. This requires to separate Symbol and a new type dedicated to the parser
because it would overload Symbol with specific parser information (as Exit already
does).