stop calling the listener once an error occurred; call final abort instead
When the parser is trying to recover, there is no guarantee that the listener
is called with the expected pattern. There's a high risk that it causes the
listener to corrupt its data or to lack expected data on the stack.

The parsing only continues to detect other possible syntax errors.