updated doc
Defining dependencies in the same workspace but with older git revision numbers is too problematic
to be considered now. The most annoying problem is that the main library at version X and at
version Y are considered as different; e.g. a type defined in that library and expected as
parameter of a function defined in a package must come from the same library version as the
one in the package code (the package being `lexi`, `make-lexi`, etc).

One solution would be to re-export the library in the package, but that defeats the purpose of
having a different version.

Another solution is to make executable binaries for packages like `make-lexi`, so we can
make changes in the lexicon or the grammar independently of the current library state.