add some tests and fixes, one remaining funny error
ERROR: syntax error: found extra symbol ':' after end of parsing

In fact, it's an error in the grammar of gram: the EOF symbol has been put
after the rules, rather that inside one of them. It will have to be an
optional element before the semicolon, and the listener will have to check
that only the top rule has it.