refactor: rename "prod" to "prule" (macro, misc fields, method names, ...)
Some of the names used are incorrect, like "factors", "ProdFactor", etc.
"A -> A a | b" is a set of two productions, "A -> A a" and "A -> b" for
the nonterminal A. "A a" and "b" are alternatives (or production
bodies), not factors.

The list of all the alternatives for one nonterminal are its production
rule. We keep it singular to avoid confusion with the set of production
rules, or production rule set, which contains all the rules of a grammar.

Thus we have the hierarchy:

- ProdRuleSet: production rule set (for all nonterminals)
  - ProdRule: production rule (for one nonterminal)
    - Alternative: alternative (or production body)

We'll rename progressively the following types and references to them in
various names: variables, parameters, method names, etc.

ProdFactor              -> Alternative

    prodf!              alt!

ProdRule                -> (same)

    prod!               prule!
    prod_to_str         prule_to_str
    prod_to_macro       prule_to_macro
    ...

ProdRuleSet             -> (same)