move rtsgen to library to avoid cyclic dependency
When a modification to Parser, ... makes it too difficult to keep
rtsgen compilable, we turn the "rtsgen" feature off. This disables
most of the tests that rely on it, so we keep a very few old tests
based on build_rts / build_prs to validate the new code just enough
before restoring rtsgen (then the other generated sources like
Lexi & Gram).